dashboard/requirements.txt
streamlit==1.45.1
plotly==6.1.1
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48
python-dotenv==1.1.1
